Remove the 1st solve button implementation
