+++ packages/scrabble-solver/src/i18n/en.json
@@ -29,6 +29,7 @@
   "results.empty-state.unitialized": "Words generated from your letters will be shown here.",
   "results.header.points": "Points",
   "results.header.word": "Word",
+  "results.solve": "Solve",
   "settings": "Settings",
   "settings.autoGroupTiles": "Group remaining tiles",
   "settings.autoGroupTiles.left": "On the left",
