+++ packages/scrabble-solver/src/i18n/pl.json
@@ -29,6 +29,7 @@
   "results.empty-state.unitialized": "Tu zostaną wyświetlone słowa wygenerowane z Twoich liter.",
   "results.header.points": "Punkty",
   "results.header.word": "Słowo",
+  "results.solve": "Rozwiąż",
   "settings": "Opcje",
   "settings.autoGroupTiles": "Grupuj pozostałe płytki",
   "settings.autoGroupTiles.left": "Po lewej",
